增加请求参数异常

+++ star-manager/src/main/java/com/rookiestar/starmanager/controller/CompanyRestController.java
@@ -9,6 +9,7 @@ import com.rookiestar.starmanager.constant.PermissionNames;
 import com.rookiestar.starmanager.constant.RoleNames;
 import com.rookiestar.starmanager.entity.position.Position;
 import com.rookiestar.starmanager.exception.HireException;
+import com.rookiestar.starmanager.exception.RequestParameterException;
 import com.rookiestar.starmanager.rabbit.MessageProducer;
 import com.rookiestar.starmanager.service.*;
 import com.rookiestar.starmanager.util.DateUtil;
@@ -83,7 +84,11 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping("/hireEmployee.do")
-    public Experience hireEmployee(int accountNumber,int departmentId,int positionId) throws Exception{
+    public Experience hireEmployee(Integer accountNumber,Integer departmentId,Integer positionId) throws Exception{
+        if(accountNumber==null||departmentId==null||positionId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
+
         Employee employee = retrieveService.retrieveEmployeeByAccountNumber(accountNumber);
         if(employee==null){
             throw new HireException("员工尚未注册");
@@ -120,6 +125,10 @@ public class CompanyRestController {
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping("/registerEmployee.do")
     public Employee registerEmployee(String name, String birthday, String gender, String email, String identifyNumber, String password) throws Exception{
+        if(name==null||birthday==null||gender==null||email==null||identifyNumber==null||password==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
+
         Employee employee = new Employee(name, DateUtil.parse(birthday),gender,email,identifyNumber,0,password,null);
         Employee newEmployee = createService.registerEmployee(employee);
 
@@ -143,6 +152,9 @@ public class CompanyRestController {
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping("/getEmployeesByName.do")
     public List<Employee> getEmployeesByName(String name){
+        if(name==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeesByName(name);
     }
 
@@ -156,6 +168,9 @@ public class CompanyRestController {
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping("/getEmployeeByIdentifyNumber.do")
     public Employee getEmployeeByIdentifyNumber(String identifyNumber){
+        if(identifyNumber==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeeByIdentifyNumber(identifyNumber);
     }
 
@@ -169,6 +184,9 @@ public class CompanyRestController {
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping("/getEmployeesByGender.do")
     public List<Employee> getEmployeesByGender(String gender){
+        if(gender==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeesByGender(gender);
     }
 
@@ -182,6 +200,9 @@ public class CompanyRestController {
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping(value = "/getEmployeeByEmail.do")
     public Employee getEmployeeByEmail(String email){
+        if(email==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeeByEmail(email);
     }
 
@@ -195,7 +216,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/updateAssessment.do")
-    public boolean updateAssessment(int accountNumber, int companyId, String startTime, String absenteeismRate, String performance) throws Exception {
+    public boolean updateAssessment(Integer accountNumber, Integer companyId, String startTime, String absenteeismRate, String performance) throws Exception {
+        if(accountNumber==null||companyId==null||startTime==null||absenteeismRate==null||performance==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         Assessment assessment = new Assessment(accountNumber, companyId, DateUtil.parse(startTime), absenteeismRate, performance);
         return updateService.updateAssessment(assessment);
     }
@@ -209,9 +233,12 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/updateExperience.do")
-    public boolean updateExperience(int accountNumber, int companyId,
-                                    int departmentId, int positionId, int jobNumber,
-                                    String startTime, String endTime, boolean isEnd) throws Exception {
+    public boolean updateExperience(Integer accountNumber, Integer companyId,
+                                    Integer departmentId, Integer positionId, Integer jobNumber,
+                                    String startTime, String endTime, Boolean isEnd) throws Exception {
+        if(accountNumber==null||companyId==null||departmentId==null||positionId==null||jobNumber==null||startTime==null||endTime==null||isEnd==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         Experience experience = new Experience(accountNumber, companyId, departmentId, positionId, jobNumber,
                 DateUtil.parse(startTime), DateUtil.parse(endTime), isEnd);
         return updateService.updateExperience(experience);
@@ -226,7 +253,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/updateDepartment.do")
-    public boolean updateDepartment(int companyId,int departmentId,String name){
+    public boolean updateDepartment(Integer companyId,Integer departmentId,String name){
+        if(companyId==null||departmentId==null||name==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         Department department = new Department(companyId, departmentId, name, null);
         return updateService.updateDepartment(department);
     }
@@ -240,7 +270,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/updatePosition.do")
-    public boolean updatePosition(int companyId,int departmentId,int positionId,String name){
+    public boolean updatePosition(Integer companyId,Integer departmentId,Integer positionId,String name){
+        if(companyId==null||departmentId==null||positionId==null||name==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         Position position = new Position(companyId, departmentId, positionId, name);
         return updateService.updatePosition(position);
     }
@@ -254,7 +287,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/deletePosition.do")
-    public boolean deletePosition(int companyId,int departmentId,int positionId){
+    public boolean deletePosition(Integer companyId,Integer departmentId,Integer positionId){
+        if(companyId==null||departmentId==null||positionId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return deleteService.deletePositionByCompanyIdAndDepartmentIdAndPositionId(companyId,departmentId,positionId);
     }
 
@@ -267,7 +303,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.WRITE})
     @RequestMapping(value = "/deleteDepartment.do")
-    public boolean deleteDepartment(int companyId,int departmentId){
+    public boolean deleteDepartment(Integer companyId,Integer departmentId){
+        if(companyId==null||departmentId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return deleteService.deleteDepartmentByCompanyIdAndDepartmentId(companyId,departmentId);
     }
 
@@ -280,7 +319,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping(value = "/getDepartmentByCompanyId.do")
-    public List<Department> getDepartmentByCompanyId(int companyId){
+    public List<Department> getDepartmentByCompanyId(Integer companyId){
+        if(companyId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveDepartmentByCompanyId(companyId);
     }
 
@@ -293,7 +335,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping(value = "/getPositionByCompanyIdAndDepartmentId.do")
-    public List<Position> getPositionByCompanyIdAndDepartmentId(int companyId,int departmentId){
+    public List<Position> getPositionByCompanyIdAndDepartmentId(Integer companyId,Integer departmentId){
+        if(companyId==null||departmentId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrievePositionByCompanyIdAndDepartmentId(companyId,departmentId);
     }
 
@@ -307,7 +352,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping(value = "/getEmployeeByCompanyIdAndDepartmentId.do")
-    public List<Employee> getEmployeeByCompanyIdAndDepartmentId(int companyId,int departmentId){
+    public List<Employee> getEmployeeByCompanyIdAndDepartmentId(Integer companyId,Integer departmentId){
+        if(companyId==null||departmentId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeesByCompanyIdAndDepartmentId(companyId,departmentId);
     }
 
@@ -320,7 +368,10 @@ public class CompanyRestController {
     @RequiresRoles(value = {RoleNames.COMPANY_MANAGER,RoleNames.MANAGER},logical = Logical.OR)
     @RequiresPermissions(value = {PermissionNames.READ})
     @RequestMapping(value = "/getEmployeeByCompanyIdAndDepartmentIdAndPositionId.do")
-    public List<Employee> getEmployeeByCompanyIdAndDepartmentIdAndPositionId(int companyId,int departmentId,int positionId){
+    public List<Employee> getEmployeeByCompanyIdAndDepartmentIdAndPositionId(Integer companyId,Integer departmentId,Integer positionId){
+        if(companyId==null||departmentId==null||positionId==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         return retrieveService.retrieveEmployeesByCompanyIdAndDepartmentIdAndPositionId(companyId,departmentId,positionId);
     }
